+++ src/services/translation/Export_ImportTranslationService.php
@@ -96,7 +96,7 @@ class Export_ImportTranslationService implements TranslationServiceInterface
 
             //$jobFactory->dispatchJob(UpdateDraftFromXml::class, $element, $draft, $target, $file->sourceSite, $file->targetSite);
 
-            Craft::$app->queue->push(new UpdateDraftFromXmlJob([
+            Craft::$app->queue->push(new UpdateDraftFromXml([
                 'description' => 'Updating Entry Drafts',
                 'element' => $element,
                 'draft' => $draft,
